lab08/src/Questions/Chapter9/Question2.java
package Questions.Chapter9;

/**
 *
 * @author (Your Name Here)
 */
public class Question2 {

    /*  For this exercise, you will be instantiating multidimensional arrays based on the data given.
    *  Once you have instantiated these arrays, you will need to store them as elements in an
    *   Object array that will be returned by the method.
     */
    public Object[] Question2() {
        /* 1) A sports columnist at the paper wants to know how many downs each member of the Bison Football
		 *	  team played during the regular season. You should construct a two-dimensional array where the
		 *	  number of rows represent the 111 players on the team's roster and the columns represent
		 *	  the 11 games of the regular season.
         */
        int[][] downsPlayed = new int[111][11];


        /* 2) One of your friends majoring in Geology is interested in storing and processing data related
		 *    to the average snowfall in North Dakota over the past 100 years. Instantiate a 2D array that
		 *    can handle floating-point numbers where the rows are the previous 100 years and the columns are
		 *    the months (November - May).
         */
        double[][] averageSnowfall = new double[100][7];


        /* 3) Your cousin has started a side-business selling ice cream cones and wants to start tracking the
		 *	  popularity of her different flavors at a series of different locations. Create a multi-dimensional
		 *	  array that will be able to track sales for each of her three locations for each of the 52 weeks in
		 *    a year for each of the 17 flavors of ice cream that she sells from the locations.
         */
        int[][][] sales = new int[3][52][17];


        /* 4) Mr. Radermacher's friend's brother Daz is convinced that the government is out to get him and has
		 *	  devised a plan to change his name to something new every day. Create a multi-dimensional array that
		 *	  will help him store a new name for every day of year for the next 3 years. Daz is convinced that
		 *    government spy planes don't work on leap days so we don't have to worry about leap years.
         */
        String[][] newNames = new String[3][365];


        /* 5) You want to impress your friends by creating a multidimensional array for a local grocery store
		 *	  chain with their sales data that can be used for tax purposes. You'll want to store the total daily
		 *	  sales (in dollars) for each cash register for the past 10 years, for every week of those years (of which
		 *    there are 52) and day of those weeks, for each of 10 stores in the chain, which all contain 12 check-out
		 *	  aisles that have an individual cash register that holds the money.
         */
        double[][][][][] salesData = new double[10][52][7][10][12];


        /* 6) Create a single-dimensional array of type Object and add each of the arrays that you've instantiated
		 *	 to that array in the order in which they were created base on the instructions above.
		 *	 This array should be returned by the method.
         */
        Object[] arrays = new Object[5];
        arrays[0] = downsPlayed;
        arrays[1] = averageSnowfall;
        arrays[2] = sales;
        arrays[3] = newNames;
        arrays[4] = salesData;

        return arrays;
    }

}
